Moved EVM config to environment.rs
